requirements.txt
kafka-python
beautifulsoup4
google
requests
selenium